[DS-1144] Maven Project Consolidation : Consolidate SWORD Client projects
