+++ dspace-sword-client/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <packaging>pom</packaging>
+    <packaging>jar</packaging>
     <groupId>org.dspace</groupId>
     <artifactId>dspace-sword-client</artifactId>
     <version>3.0-SNAPSHOT</version>
@@ -13,16 +13,28 @@
         <version>3.0-SNAPSHOT</version>
         <relativePath>..</relativePath>
     </parent>
-    
-    <modules>
-        <module>dspace-sword-client-api</module>
-        <module>dspace-sword-client-xmlui-api</module>
-        <module>dspace-sword-client-xmlui-webapp</module>
-    </modules>
 
-    <issueManagement>
-      <system>JIRA</system>
-      <url>http://jira.dspace.org/jira/browse/SWORD</url>
-    </issueManagement>
-    
+   <dependencies>
+
+      <dependency>
+         <groupId>org.dspace</groupId>
+         <artifactId>dspace-api</artifactId>
+      </dependency>
+
+      <dependency>
+         <groupId>org.dspace</groupId>
+         <artifactId>dspace-sword</artifactId>
+         <type>jar</type>
+         <classifier>classes</classifier>
+      </dependency>
+
+      <!-- external -->
+      <dependency>
+         <groupId>javax.servlet</groupId>
+         <artifactId>servlet-api</artifactId>
+         <scope>provided</scope>
+      </dependency>
+
+   </dependencies>
+
 </project>
