+++ dspace-sword-client/src/main/java/org/dspace/sword/client/exceptions/HttpException.java
