+++ dspace-sword/pom.xml
@@ -26,7 +26,6 @@
             <artifactId>maven-war-plugin</artifactId>
             <configuration>
                <attachClasses>true</attachClasses>
-               <archiveClasses>true</archiveClasses>
                <!-- In version 2.1-alpha-1, this was incorrectly named warSourceExcludes -->
                <packagingExcludes>WEB-INF/lib/*.jar</packagingExcludes>
                <warSourceExcludes>WEB-INF/lib/*.jar</warSourceExcludes>
