+++ dspace-xmlui/pom.xml
@@ -159,6 +159,13 @@
          <artifactId>dspace-discovery-solr</artifactId>
       </dependency>
 
+      <dependency>
+         <groupId>org.dspace</groupId>
+         <artifactId>dspace-sword-client</artifactId>
+
+      </dependency>
+
+
       <!-- Custom build DSpace cocoon -->
       <dependency>
          <groupId>org.apache.cocoon</groupId>
