+++ dspace-xmlui/src/main/resources/aspects/SwordClient/sitemap.xmap
