+++ dspace/modules/xmlui/pom.xml
@@ -135,19 +135,6 @@
          <type>war</type>
       </dependency>
 
-      <dependency>
-         <groupId>org.dspace</groupId>
-         <artifactId>dspace-sword-client-xmlui-api</artifactId>
-         <version>3.0-SNAPSHOT</version>
-      </dependency>
-
-       <dependency>
-         <groupId>org.dspace</groupId>
-         <artifactId>dspace-sword-client-xmlui-webapp</artifactId>
-         <version>3.0-SNAPSHOT</version>
-         <type>war</type>
-      </dependency>
-
       <!-- DSpace XMLUI API -->
       <dependency>
          <groupId>org.dspace</groupId>
@@ -162,14 +149,6 @@
          <version>2.3</version>
          <scope>provided</scope>
       </dependency>
-
-       <!--  A bit untidy but needed for XMLUI Sword client -->
-       <dependency>
-           <groupId>org.dspace</groupId>
-           <artifactId>dspace-sword</artifactId>
-           <type>jar</type>
-           <classifier>classes</classifier>
-       </dependency>
 
    </dependencies>
